refactor: Create service layer and remove unused imports
