--- src/lambdas/in.json
@@ -1,6 +1,0 @@
-{
-    "body": {
-        "username": "admin",
-        "password": "admin"
-    }
-}
