+++ src/lambdas/lambda_login.py
@@ -18,14 +18,12 @@ SECRET = "verysecret"
 session = boto3.Session(aws_access_key_id=ACCESS_KEY, aws_secret_access_key=SECRET_KEY)
 dynamo_cli = session.client('dynamodb', endpoint_url=ENDPOINT)
 
-
 """
 awslocal lambda create-function --function-name login --zip-file fileb://login.zip --runtime python3.9 --handler lambda_login.lambda_login --role arn:aws:iam::000000000000:role/LambdaBasic
 awslocal lambda update-function-configuration --function-name login --timeout 3
 awslocal lambda update-function-code --function-name login --zip-file fileb://login.zip
 awslocal lambda invoke --function-name login  --payload file://in.json ./out.json
 """
-
 
 # https://stackoverflow.com/a/68409773
