--- src/lambdas/out.json
@@ -1,1 +1,0 @@
-{"body": {"token": "eyJhbGciOiAiSFMyNTYiLCAidHlwIjogIkpXVCJ9.eyJ1c2VybmFtZSI6ICJhZG1pbiJ9.ZTc5NjQ1ZDMwM2NjYTlkMWZlNmI5YjU5ODNiNzc1MmY0NGJkYjE3M2E1MzllMjZiYTgwOTdiODRjZDYzNDk2Nw"}}
